+++ test/killer_sudoku.jl
@@ -124,18 +124,19 @@ end
         end
         com = JuMP.backend(m).optimizer.model.inner
         @test general_tree_test(com)
-       
+
         return com
     end
 
     @testset "Killer Sudoku niallsudoku_5503 with negative coefficients and -9 to -1" begin
         com1 = killer_negative()
-        @test_reference "refs/niallsudoku_5503_negative" test_string([constraint.indices for constraint in com1.constraints])
+        com1_str = test_string([constraint.indices for constraint in com1.constraints])
         # the constraint order now affects this but calling the same twice should have the same ordering
         # this tests that the order of constraints the user specified is actually moved correctly to ConstraintSolver instead of being random
         # see https://github.com/Wikunia/ConstraintSolver.jl/issues/180
         com2 = killer_negative()
-        @test_reference "refs/niallsudoku_5503_negative" test_string([constraint.indices for constraint in com2.constraints])
+        com2_str = test_string([constraint.indices for constraint in com2.constraints])
+        @test com1_str == com2_str
         info_1 = com1.info
         info_2 = com2.info
         @test info_1.pre_backtrack_calls == info_2.pre_backtrack_calls